requirements.txt
keras==2.8.0
numpy==1.21.4
opencv-python-headless<4.5.5.64
Pillow==9.1.0
streamlit==1.2.0


